fix: 폴더 분리 & rem 값 조정
