+++ src/components/SettingPage/ProfileArea.tsx
@@ -1,9 +1,6 @@
 import styled from '@emotion/styled';
 
-import google_calendar from '@/assets/images/google_calendar.png';
 import sampleImg from '@/assets/images/sample.png';
-import Icons from '@/assets/svg/index';
-import SettingXBtn from '@/components/common/button/settingBtn/SettingXBtn';
 import USERS from '@/constants/users';
 
 function ProfileArea() {
@@ -17,20 +14,6 @@ function ProfileArea() {
 					<EmailText>{USERS.data.email}</EmailText>
 				</ProfileTextBox>
 			</ProfileWrapper>
-			<AccountText>연동된 계정</AccountText>
-			<AccountWrapper>
-				<CalendarContainer>
-					<CalendarBox>
-						<GoogleImg src={google_calendar} alt="구글 캘린더" />
-						<CalendarText>구글 캘린더</CalendarText>
-					</CalendarBox>
-					<IconBox />
-				</CalendarContainer>
-				<InputBox>
-					<EmailWrapper>{USERS.data.email}</EmailWrapper>
-					<SettingXBtn />
-				</InputBox>
-			</AccountWrapper>
 		</ProfileAreaLayout>
 	);
 }
@@ -42,6 +25,7 @@ const ProfileAreaLayout = styled.div`
 	flex-direction: column;
 	align-items: flex-start;
 	max-width: 43.2rem;
+	height: 17.8rem;
 `;
 
 const ProfileWrapper = styled.div`
@@ -85,80 +69,4 @@ const ProfileText = styled.div`
 	align-items: center;
 	${({ theme }) => theme.fontTheme.HEADLINE_02}
 	color: ${({ theme }) => theme.palette.Grey.Black}
-`;
-
-const AccountText = styled.div`
-	display: flex;
-	padding: 3.6rem 3.6rem 0.8rem 2rem;
-	justify-content: center;
-	align-items: center;
-	${({ theme }) => theme.fontTheme.HEADLINE_02}
-	color: ${({ theme }) => theme.palette.Grey.Black}
-`;
-
-const AccountWrapper = styled.div`
-	display: flex;
-	width: 412px;
-	padding: 1.2rem 1.6rem 1.6rem 1.6rem;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	gap: 1.2rem;
-	margin: 0.8rem 1rem 0.8rem 0.7rem;
-	border-radius: 12px;
-	border: 1px solid #e9e9ee;
-	box-sizing: border-box;
-`;
-
-const CalendarContainer = styled.div`
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	align-self: stretch;
-	width: 38rem;
-	height: 4rem;
-`;
-
-const CalendarBox = styled.div`
-	height: 4rem;
-	display: flex;
-	justify-content: flex-start;
-	gap: 0.4rem;
-	box-sizing: border-box;
-`;
-
-const GoogleImg = styled.img`
-	margin: 0.2rem 0;
-	width: 3.6rem;
-	height: 3.6rem;
-`;
-
-const CalendarText = styled.p`
-	${({ theme }) => theme.fontTheme.BODY_02};
-	width: 7.4rem;
-	align-self: center;
-	margin: 0.6rem 0;
-`;
-
-const IconBox = styled(Icons.plus_circle)`
-	width: 2.4rem;
-	height: 2.4rem;
-	margin: 0.8rem;
-	box-sizing: border-box;
-`;
-
-const InputBox = styled.p`
-	display: flex;
-	width: 38.2rem;
-	padding: 0.6rem 0.8rem 0.6rem 1.2rem;
-	justify-content: space-between;
-	align-items: center;
-	border-radius: 8px;
-	background: #dfe9fc;
-	box-sizing: border-box;
-`;
-
-const EmailWrapper = styled.p`
-	${({ theme }) => theme.fontTheme.BODY_02};
-	color: #5a5a61;
 `;
